requirements.txt
-i https://pypi.org/simple
annotated-types==0.5.0 ; python_version >= '3.7'
anyio==3.7.1 ; python_version >= '3.7'
bcrypt==4.0.1
certifi==2023.7.22 ; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.2.0 ; python_full_version >= '3.7.0'
click==8.1.7 ; python_version >= '3.7'
cryptography==41.0.3
dnspython==2.4.2 ; python_version >= '3.8' and python_version < '4.0'
ecdsa==0.18.0 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
email-validator==2.0.0.post2
fastapi[all]==0.101.1
h11==0.14.0 ; python_version >= '3.7'
httpcore==0.17.3 ; python_version >= '3.7'
httptools==0.6.0
httpx==0.24.1
idna==3.4 ; python_version >= '3.5'
iniconfig==2.0.0 ; python_version >= '3.7'
itsdangerous==2.1.2
jinja2==3.1.2
markupsafe==2.1.3 ; python_version >= '3.7'
orjson==3.9.5
packaging==23.1 ; python_version >= '3.7'
passlib[bcrypt]==1.7.4
pluggy==1.2.0 ; python_version >= '3.7'
pyasn1==0.5.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycparser==2.21
pydantic==2.3.0 ; python_version >= '3.7'
pydantic-core==2.6.3 ; python_version >= '3.7'
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pymongo==4.5.0
pytest==7.4.0
python-dotenv==1.0.0 ; python_version >= '3.8'
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pyyaml==6.0.1
requests==2.31.0
rsa==4.9 ; python_version >= '3.6' and python_version < '4'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
starlette==0.27.0 ; python_version >= '3.7'
typing-extensions==4.7.1 ; python_version >= '3.7'
ujson==5.8.0
urllib3==2.0.4 ; python_version >= '3.7'
uvicorn[standard]==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
